[Update] comment 컴포넌트 생성 (post 컴포넌트 분리)
